Extract Monster::Base class
