+++ lib/thirteenth_age/monster/large.rb
@@ -1,42 +1,19 @@
 module ThirteenthAge
   module Monster
-    class Large
-      STRIKE_DAMAGE_PER_LEVEL = [9,10,14,21,28,36,42,56,76,100,116,140,180,220,270]
-      HP_PER_LEVEL = [41,54,72,90,108,144,180,216,288,360,432,576,720,864,1152]
+    class Large < Base
 
-      attr_accessor :level
-
-      def initialize(options = {})
-        self.level = options[:level]
+      def strike_damage_per_level
+        [9,10,14,21,28,36,42,56,76,100,116,140,180,220,270]
       end
 
-      def attack_bonus
-        level + 5
+      def hp_per_level
+        [41,54,72,90,108,144,180,216,288,360,432,576,720,864,1152]
       end
 
-      def strike_damage
-        STRIKE_DAMAGE_PER_LEVEL[level]
+      def fear_factor
+        6.0
       end
 
-      def hp
-        HP_PER_LEVEL[level]
-      end
-
-      def ac
-        level + 16
-      end
-
-      def better_defense
-        level + 14
-      end
-
-      def lesser_defense
-        level + 10
-      end
-
-      def fear_threshold
-        (hp / 6.0).ceil
-      end
     end
   end
 end
