+++ lib/thirteenth_age/monster/mook.rb
@@ -1,42 +1,19 @@
 module ThirteenthAge
   module Monster
-    class Mook
-      STRIKE_DAMAGE_PER_LEVEL = [3,4,5,6,7,9,12,18,23,31,37,46,60,74,90]
-      HP_PER_LEVEL = [5,7,9,11,14,18,23,27,36,45,54,72,90,108,144]
+    class Mook < Base
 
-      attr_accessor :level
-
-      def initialize(options = {})
-        self.level = options[:level]
+      def strike_damage_per_level
+        [3,4,5,6,7,9,12,18,23,31,37,46,60,74,90]
       end
 
-      def attack_bonus
-        level + 5
-      end
-
-      def strike_damage
-        STRIKE_DAMAGE_PER_LEVEL[level]
-      end
-
-      def hp
-        HP_PER_LEVEL[level]
-      end
-
-      def ac
-        level + 16
-      end
-
-      def better_defense
-        level + 14
-      end
-
-      def lesser_defense
-        level + 10
+      def hp_per_level
+        [5,7,9,11,14,18,23,27,36,45,54,72,90,108,144]
       end
 
       def fear_threshold
         nil
       end
+
     end
   end
 end
